+++ geotrek/api/tests/test_v2.py
@@ -2110,6 +2110,37 @@ class APIAccessAnonymousTestCase(BaseApiTest):
             ORGANISM_PROPERTIES_JSON_STRUCTURE
         )
 
+    def test_sensitivearea_distance_list(self):
+        p1 = core_factory.PathFactory.create(geom=LineString((605600, 6650000), (605604, 6650004), srid=2154))
+        trek = trek_factory.TrekFactory.create(
+            published=True,
+            name='Parent',
+            paths=[p1]
+        )
+        specy = sensitivity_factory.SpeciesFactory.create(period01=True)
+        sensitivity_factory.SensitiveAreaFactory.create(
+            geom='SRID=2154;POLYGON((605600 6650000, 605600 6650004, 605604 6650004, 605604 6650000, 605600 6650000))',
+            species=specy,
+            description="Test"
+        )
+        sensitivity_factory.SensitiveAreaFactory.create(
+            geom='SRID=2154;POLYGON((606001 6650501, 606001 6650505, 606005 6650505, 606005 6650501, 606001 6650501))',
+            species=specy,
+            description="Test 2"
+        )
+        response = self.get_sensitivearea_list({
+            'trek': trek.id,
+            'period': '1'
+        })
+        self.assertEqual(response.json()['count'], 1)
+        self.assertEqual(response.status_code, 200)
+        with override_settings(SENSITIVE_AREA_INTERSECTION_MARGIN=700):
+            response = self.get_sensitivearea_list({
+                'trek': trek.id,
+                'period': '1'
+            })
+            self.assertEqual(response.json()['count'], 2)
+
 
 class APIAccessAdministratorTestCase(BaseApiTest):
     """